# BackendV2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FUTCOL API",
        "description": "API para el sistema de gestión de torneos de fútbol FUTCOL. Esta API permite gestionar torneos, equipos, jugadores, amonestaciones y estadísticas.",
        "contact": {
            "name": "FUTCOL Support Team",
            "email": "admin@futcol.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "FUTCOL API Server"
        }
    ],
    "paths": {
...
        "/api/sedes": {
            "get": {
                "tags": [
                    "Sedes"
                ],
                "summary": "Listar todas las sedes",
                "operationId": "957205a3296ece9539db8919256438c0",
                "responses": {
                    "200": {
                        "description": "Lista de sedes con su torneo asociado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sedes"
                ],
                "summary": "Crear una nueva sede",
                "operationId": "5bfd771571873bfafbe26f64d6cc4285",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion",
                                    "torneo_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Cancha el Dorado"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Carrera 12 #45-67"
                                    },
                                    "torneo_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sede creada exitosamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/sedes/{id}": {
            "get": {
                "tags": [
                    "Sedes"
                ],
                "summary": "Mostrar una sede específica",
                "operationId": "82761c19368834a3a53044485b94a9a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información de la sede"
                    },
                    "404": {
                        "description": "Sede no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sedes"
                ],
                "summary": "Actualizar una sede",
                "operationId": "b53c2fdfd7e19e2be4e529c19606422f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nuevo nombre de sede"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Nueva dirección"
                                    },
                                    "torneo_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sede actualizada correctamente"
                    },
                    "404": {
                        "description": "Sede no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sedes"
                ],
                "summary": "Eliminar una sede",
                "operationId": "5cf9daa3fe988bf044a1c65785bf9e9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sede eliminada con éxito"
                    },
                    "404": {
                        "description": "Sede no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT en el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}